# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Contact Admin System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id?: number | null;
  telegram_id: string;
  first_name?: string | null;
  username?: string | null;
  is_admin: boolean;
}

export type MessageSenderType =
  (typeof MessageSenderType)[keyof typeof MessageSenderType];

export const MessageSenderType = {
  user: "user",
  admin: "admin",
} as const;

export interface Message {
  id: number;
  user_id: number;
  sender_type: MessageSenderType;
  text?: string | null;
  media_type: string;
  media_url?: string | null;
  telegram_file_id?: string | null;
  created_at: string;
}

export interface UserWithLastMessage {
  id: number;
  telegram_id: string;
  first_name?: string | null;
  username?: string | null;
  last_msg?: string | null;
  last_msg_at?: string | null;
  last_media_type?: string | null;
}

export interface SendMessageRequest {
  text: string;
  targetUserId?: number | null;
}

export interface BroadcastRequest {
  text: string;
}

export interface BroadcastResponse {
  ok: boolean;
  sent: number;
  total: number;
}

export interface OkResponse {
  ok: boolean;
}

export interface CreateDonationRequest {
  amount: number;
  currency?: string;
}

export interface DonationInvoice {
  ok: boolean;
  trackId?: string;
  payLink?: string | null;
  qrImage?: string | null;
}

export type DonationStatusDonation = { [key: string]: unknown } | null;

export interface DonationStatus {
  ok: boolean;
  oxaStatus?: string;
  donation?: DonationStatusDonation;
}

export interface Donation {
  id: number;
  user_id: number;
  amount: number;
  status: string;
  tx_id?: string | null;
  track_id?: string | null;
  created_at: string;
}