# lib/api-zod/src/generated/types/createDonationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Contact Admin System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDonationRequest {
  amount: number;
  currency?: string;
}