# lib/api-zod/src/generated/types/donation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Contact Admin System API
 * OpenAPI spec version: 0.1.0
 */

export interface Donation {
  id: number;
  user_id: number;
  amount: number;
  status: string;
  tx_id?: string | null;
  track_id?: string | null;
  created_at: string;
}