# lib/api-zod/src/generated/types/donationStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Contact Admin System API
 * OpenAPI spec version: 0.1.0
 */
import type { DonationStatusDonation } from "./donationStatusDonation";

export interface DonationStatus {
  ok: boolean;
  oxaStatus?: string;
  donation?: DonationStatusDonation;
}