# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Contact Admin System API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageSenderType } from "./messageSenderType";

export interface Message {
  id: number;
  user_id: number;
  sender_type: MessageSenderType;
  text?: string | null;
  media_type: string;
  media_url?: string | null;
  telegram_file_id?: string | null;
  created_at: string;
}