# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Contact Admin System API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id?: number | null;
  telegram_id: string;
  first_name?: string | null;
  username?: string | null;
  is_admin: boolean;
}