# lib/api-zod/src/generated/types/userWithLastMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Contact Admin System API
 * OpenAPI spec version: 0.1.0
 */

export interface UserWithLastMessage {
  id: number;
  telegram_id: string;
  first_name?: string | null;
  username?: string | null;
  last_msg?: string | null;
  last_msg_at?: string | null;
  last_media_type?: string | null;
}